Fix incorrect commandline argument
The way to run in cpu mode is with --gpu_ids -1. Likely it was changed at some point and the docs need to be updated to reflect that change. --device_ids -1 isn't accepted as a commanline argument to python train.py.

Apologies if the --device_ids -1 is a correct commandline argument to a different .py file!